31_Hospital_management/backend/requirements.txt
asgiref==3.11.1
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
ConfigArgParse==1.7.5
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.5.0
h11==0.16.0
idna==3.13
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.4
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.2
pluggy==1.6.0
psutil==7.2.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
python-engineio==4.13.1
python-socketio==5.16.1
pywin32==311
pyzmq==27.1.0
requests==2.33.1
simple-websocket==1.1.0
sqlparse==0.5.5
tzdata==2026.2
urllib3==2.6.3
websocket-client==1.9.0
Werkzeug==3.1.8
wsproto==1.3.2
zope.event==6.2
zope.interface==8.4
